fix some generic constraints
this is in fact pretty interesting. even though a generic was defined as `C: Component`, the compiler would "limit" it to `C: Handler`, because the C was used in some other parameter type which is constrained to Handler. The xcode-generated docs show the correct (ie, compiler-defined) type constraints but there's no indication whatsoever as to the fact that this change happened